Rename service to better suggest object origin
